[Refactor] Move ScriptLib function implementation to scriptlib package
